PentestingFundamentals
Just finished the Penetration Testing Fundamentals room on TryHackMe — and I really enjoyed this one! It covered the core methodologies like OWASP, NIST, OSSTMM, and NCSC CAF, and walked through the process of identifying and exploiting vulnerabilities in a structured and ethical way.

I especially liked learning how to scan for vulnerabilities on a web server, understand attacker workflows, and see how real-world assessments are conducted.