HackingHTML webpage
I enjoyed this, because it helped utilize basic commands to see how easily it was to access what would seem to an ordinary user "invisible". 
Going into the HTML revealed a lot, comments, files and other information that quite simply should not have been there. 
If I have learnt anything from this experience, it is ATTENTION TO DETAIL. 100%! Is one of the most important elements at play here.