THM - NmapBasics
Learning how to scan networks and enumerate services is such a fundamental step in ethical hacking and cyber security. Nmap is an incredibly versatile tool from simple host discovery to advanced scripting for vulnerability detection, it’s an essential skill for both red and blue teamers.

This course helped me solidify the foundations and gave me more confidence in using Nmap for practical scenarios. Excited to keep building on this knowledge as I progress further!